Fix freeing wasmtime data structures on multiple threads
The C API is not threadsafe right now and structures cannot be dropped
concurrently on other threads. Unfortunately though with `SetFinalizer`
that spins up the finalizer in an arbitrary goroutine which can run on a
foreign thread, and this appears to be the cause of crashes in CI.

Instead what we need to do is to have all finalizers run on the same
thread performing other wasmtime operations. The best approximation of
this that I can think of is to have finalizers enqueue pointers to get
deleted, and then when the main wasmtime thread is doing something it
occasionally checks to see if it should free internal objects.

Not exactly a great solution for now, but it should fix a memory safety
issue at least!