Update with Wasmtime's new C API (#81)
* Update with Wasmtime's new C API

This commit updates this repository to use Wasmtime's new C API as
specified by [RFC 11]. This is a breaking change for this API and
notably requires that a `Storelike` object is now passed in for many
functions, serving as the `Store` context for that method. The
implementation of memory management, however, has been greatly
simplified, now that objects within a `Store` no longer need
destructors. This should improve performance and also help avoid
reclaiming memory only at odd times.

[RFC 11]: https://github.com/bytecodealliance/rfcs/pull/11

* Try fixing bazel build

* Cancel in-flight jobs on CI